requirements.txt
aiogram>=3.20
aiosqlite
aiohttp
google-genai
python-dotenv
